refactor: :recycle: refactor database session and cache code
